usuario 0.9.2
se agregaron validaciones y excepciones

+++ usuario/src/main/java/com/medilink/usuario/Service/UsuarioService.java
@@ -2,8 +2,11 @@ package com.medilink.usuario.Service;
 
 import com.medilink.usuario.dto.request.UsuarioRequest;
 import com.medilink.usuario.dto.response.UsuarioResponse;
+import com.medilink.usuario.exception.CorreoDuplicadoException;
 import com.medilink.usuario.exception.UsuarioNoEncontradoExeption;
 import com.medilink.usuario.mapper.UsuarioMapper;
+import com.medilink.usuario.modelo.EstadoUsuario;
+import com.medilink.usuario.modelo.RolUsuario;
 import com.medilink.usuario.modelo.Usuario;
 import com.medilink.usuario.repository.UsuarioRepository;
 import lombok.RequiredArgsConstructor;
@@ -29,10 +32,20 @@ public class  UsuarioService {
      }
 
      public UsuarioResponse crear(UsuarioRequest usuarioRequest){
-        return usuarioMapper.toResponse(usuarioRepository.save(usuarioMapper.toEntity(usuarioRequest)));
-     }
 
+         if (usuarioRepository.existsByCorreoUsuario(usuarioRequest.getCorreoUsuario())) {
+             throw new CorreoDuplicadoException(usuarioRequest.getCorreoUsuario()
+             );
+         }
+
+         return usuarioMapper.toResponse(usuarioRepository.save(usuarioMapper.toEntity(usuarioRequest)));
+     }
      public void eliminar(Long idBuscado){
+
+        if (!usuarioRepository.existsById(idBuscado)) {
+            throw new UsuarioNoEncontradoExeption(idBuscado);
+        }
+
         usuarioRepository.deleteById(idBuscado);
      }
 
@@ -44,8 +57,9 @@ public class  UsuarioService {
         usuarioExistente.setNombreUsuario(usuarioRequest.getNombreUsuario());
         usuarioExistente.setCorreoUsuario(usuarioRequest.getCorreoUsuario());
         usuarioExistente.setContrasennaUsuario(usuarioRequest.getContrasennaUsuario());
-        usuarioExistente.setRolUsuario(usuarioRequest.getRolUsuario());
-        usuarioExistente.setEstadoUsuario(usuarioRequest.getEstadoUsuario());
+
+         usuarioExistente.setRolUsuario(RolUsuario.fromValor(usuarioRequest.getRolUsuario()));
+         usuarioExistente.setEstadoUsuario(EstadoUsuario.fromValor(usuarioRequest.getEstadoUsuario()));
 
         return usuarioMapper.toResponse(usuarioRepository.save(usuarioExistente));
